+++ packages/moonstone/VirtualList/VirtualListBaseNative.js
@@ -214,6 +214,14 @@ class VirtualListCoreNative extends Component {
 		}
 	}
 
+	shouldComponentUpdate (nextProps, nextState) {
+		if ((this.props.dataSize !== nextProps.dataSize) &&
+			(nextState.firstIndex + nextState.numOfItems) < nextProps.dataSize) {
+			return false;
+		}
+		return true;
+	}
+
 	scrollBounds = {
 		clientWidth: 0,
 		clientHeight: 0,